test: split CLI diagnostics tests
